Clean the historical import, and prove the SQL port against real data
The parity gate now runs end to end on the exported Sheet, and the headline is
that it passes where it matters:

    SQL vs oracle (is the port faithful?)
      SQL <-> oracle: identical        <- over all 1,171 real rows

That was the migration's main risk. The SQL translation of computePercentiles()
reproduces the Apps Script exactly on production data, not just on fixtures.

On top of that, three product decisions from the user about the historical set,
recorded in full in docs/agent/DECISIONS.md. The old Google Form took salary as
free text with no validation, so a value that looks wrong usually is wrong:

  - Portugal only (604 of 742 rows). The Compass is a Portuguese benchmark.
  - Drop 27 rows that cannot be trusted: 18 above EUR 200k and 9 that are
    monthly pay quoted in an annual field. The user chose to remove rather than
    guess a replacement.
  - Repair 6 rows below 100, where "18" plainly meant 18,000.

The repair boundary is 100, not 1000, and that is the user catching an error I
had already written down: the row reading 450 has 48000 in its with-perks
column, so it means 45,000. A blanket x1000 would have invented a EUR 450,000
salary and published it. The with-perks total is the anchor precisely because
it is a second, independently typed figure.

Cleaning is off by default. The gate's whole value is comparing like with like,
so the default import stays raw and the three benchmarks must agree to the
euro; --clean is a separate mode that reports the impact instead of passing or
failing. Running only the clean mode would quietly retire the control.
historical.base_salary is bigint for the same reason: the clean import would
fit in an integer, but the raw control has to swallow a 3,120,000,000 entry.

Every dropped and repaired row is written to an ignored log file. Dropping rows
means the table stops mirroring the Sheet, and "we removed 27 rows, here is
exactly which" has to stay answerable.

56 tests, 9 of them new and each naming the decision and boundary it guards.

One question is left open and is recorded as the blocker in STATUS.md: oracle
vs production differs in 31 values while the row sets are identical (1030 =
1030, every bucket matching), only in the upper tail, only in historical rows.
Either the CSV export and getValues() disagree on some cells, or production is
serving a stale computation. It does not block building; it blocks claiming the
benchmark is reproduced.

+++ db/sql/001_schema.sql
@@ -142,11 +142,17 @@ create index if not exists email_log_contact_idx on email_log (contact_id, creat
 -- nothing writes here again, which is why it carries no constraints beyond
 -- what the benchmark needs. `outlier` and `country` are the two filters
 -- computePercentiles() applies before using a row.
+-- The money columns are bigint, not integer, and deliberately so. The clean
+-- import (decisions A–C) drops everything implausible and would fit in an
+-- integer comfortably — but the parity control imports the sheet RAW, and the
+-- raw sheet contains a 3 120 000 000 entry that overflows int4. Narrowing this
+-- would work right up until it silently disabled the one check that proves the
+-- SQL port reproduces production.
 create table if not exists historical (
   id            bigint generated always as identity primary key,
   country       text,
-  base_salary   integer,
-  total_comp    integer,
+  base_salary   bigint,
+  total_comp    bigint,
   role_raw      text,
   yoe_raw       text,
   outlier       boolean not null default false
